New Services

+++ backend/src/main/kotlin/com/dinnye/backend/service/interfaces/SimpleCrudService.kt
@@ -2,8 +2,8 @@ package com.dinnye.backend.service.interfaces
 
 interface SimpleCrudService<T> {
     fun create(entity: T): T
-    fun get(id: Int): T
+    fun get(id: Long): T
     fun getAll(): List<T>
     fun update(entity: T): T
-    fun delete(id: Int);
+    fun delete(id: Long);
 }
